+++ yiran-admin/yiran-admin-framework/src/main/java/one/yiran/dashboard/manage/service/impl/SysUserServiceImpl.java
@@ -2,9 +2,7 @@ package one.yiran.dashboard.manage.service.impl;
 
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.Predicate;
-import com.querydsl.core.types.SubQueryExpression;
 import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import one.yiran.dashboard.common.constants.UserConstants;
 import one.yiran.common.domain.PageModel;
@@ -18,8 +16,6 @@ import one.yiran.dashboard.vo.UserPageVO;
 import one.yiran.db.common.service.CrudBaseServiceImpl;
 import lombok.extern.slf4j.Slf4j;
 import one.yiran.dashboard.common.constants.Global;
-import one.yiran.dashboard.manage.dao.*;
-import one.yiran.dashboard.manage.entity.*;
 import one.yiran.common.exception.BusinessException;
 import one.yiran.dashboard.common.expection.user.UserNotFoundException;
 import one.yiran.dashboard.manage.service.SysConfigService;
@@ -79,19 +75,6 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
     @Override
     public SysUser findUser(Long id) {
         SysUser u = userDao.findByUserId(id);
-//        List<SysRole> roleList = roleDao.findAll();
-//        List<SysUserRole> userRoleList = userRoleDao.findAllByUserId(u.getUserId());
-//        List<Long> roleIds = userRoleList.stream().map(t -> t.getRoleId()).collect(Collectors.toList());
-//
-//        roleList.stream().forEach(t -> {
-//            if (roleIds.contains(t.getRoleId())) {
-//                t.setFlag(true);
-//            } else {
-//                t.setFlag(false);
-//            }
-//        });
-//        u.setSysRoles(roleList);
-        checkDeptModify(u);
         return u;
     }
 
@@ -99,21 +82,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
     public SysUser findUserByLoginName(String loginName) {
         Assert.notNull(loginName, "用户名不能为空");
         SysUser u = userDao.findByLoginName(loginName);
-        checkDeptModify(u);
         return u;
-    }
-
-
-    private void checkDeptModify(SysUser u) {
-//        if (u == null)
-//            return;
-//        SysDept sysDept = deptDao.findByDeptId(u.getDeptId());
-//        if (sysDept != null) {
-//            if (!StringUtils.equalsIgnoreCase(u.getDeptName(), sysDept.getDeptName())) {
-//                u.setDeptName(sysDept.getDeptName());
-//                userDao.save(u);
-//            }
-//        }
     }
 
     @Override
@@ -141,7 +110,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
     @Transactional
     @Override
     public SysUser saveUserAndPerms(SysUser user) throws BusinessException {
-        checkUserAllowed(user,"操作");
+        checkAdminModifyAllowed(user,"操作");
         saveUser(user);
         doUserPerms(user);
         return user;
@@ -261,14 +230,6 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
     @Override
     public PageModel<SysUser> getPage(PageRequest pageRequest, SysUser searchUser) {
 
-//        BooleanExpression pp = new BooleanExpression();
-//        BooleanExpression p = addSearchToQuery(searchUser);
-//
-       // PageRequestUtil.injectQuery(pageRequest,query);
-
-//        Path<String> statusPath = Expressions.stringPath("status");
-//        BooleanOperation bpre = Expressions.predicate(Ops.EQ, statusPath, Expressions.constant(request.getStatus()));
-
         List<Predicate> pres = new ArrayList<>();
         if(StringUtils.isNotBlank(searchUser.getLoginName())){
             pres.add(QSysUser.sysUser.loginName.eq(searchUser.getLoginName()));
@@ -315,8 +276,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
     public long getListSize(SysUser searchUser, Date bTime, Date eTime) {
         QSysUser qSysUser = QSysUser.sysUser;
         BooleanExpression p = qSysUser.createTime.between(bTime,eTime);
-        addSearchToQuery(searchUser, p);
-        return super.count(p);
+        return super.count(null,searchUser,p);
     }
 
     @Override
@@ -385,7 +345,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
         if(userDao.findByUserId(userId) == null ){
             throw new UserNotFoundException(userId);
         }
-        checkUserAllowed(new SysUser(userId),"删除");
+        checkAdminModifyAllowed(new SysUser(userId),"删除");
         //删除用户角色
         int delRoleCount = sysRoleService.deleteAuthUsers(userId);
         //删除用户岗位
@@ -484,7 +444,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
             try {
                 // 验证是否存在这个用户
                 SysUser u = userDao.findByLoginName(user.getLoginName());
-                checkUserAllowed(u,"导入");
+                checkAdminModifyAllowed(u,"导入");
                 if (u == null) {
                     //user.setUserId(mongoSequenceService.getNextId(SysUser.SEQUENCE_KEY));
                     user.setSalt(Global.getSalt());
@@ -561,49 +521,6 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
         }
     }
 
-    private void addSearchToQuery(SysUser searchUser, BooleanExpression p) {
-        List<BooleanExpression> criterias = buildCriterias(searchUser);
-        if (criterias.size() > 0) {
-            criterias.forEach(e -> p.and(e));
-        }
-    }
-
-
-    private void addSearchToAggregate(SysUser searchUser, List<BooleanExpression> aggregationOperations) {
-        List<BooleanExpression> criterias = buildCriterias(searchUser);
-        if (criterias.size() > 0) {
-            criterias.forEach(e -> aggregationOperations.add(e));
-        }
-    }
-
-
-    private List<BooleanExpression> buildCriterias(SysUser searchUser) {
-        List<BooleanExpression> criterias = new ArrayList<>();
-        if (searchUser != null) {
-            if (StringUtils.isNotBlank(searchUser.getLoginName())) {
-                criterias.add(QSysUser.sysUser.loginName.likeIgnoreCase("%" + searchUser.getLoginName() + "%"));
-            }
-            if (searchUser.getDeptId() != null) {
-                //criterias.add(Criteria.where("deptId").is(searchUser.getDeptId()));   ggg fvv
-//                List<Long> allDescDeptIds = null;
-//                if (searchUser.getDeptId() != null) {
-//                    allDescDeptIds = new ArrayList<>();
-//                    allDescDeptIds.add(searchUser.getDeptId());
-//                    SysDept sysDept = deptDao.findByDeptId(searchUser.getDeptId());
-//                    if (sysDept != null) {
-//                        allDescDeptIds.addAll(sysDept.getDescendents());
-//                    }
-//                }
-//                if (allDescDeptIds != null) {
-//                    //查找这个deptID下面的所有部门 的 用户
-//                    criterias.add(Criteria.where("deptId").in(allDescDeptIds));
-//                }
-            }
-        }
-        //ncriterias.add(CriteriaUtil.notDeleteCriteria());
-        return criterias;
-    }
-
     private void doUserPerms(SysUser user) {
         Long userId = user.getUserId();
         //如果传递的roleIds为空，表示不修改角色
@@ -621,72 +538,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
 
     private PageModel<SysUser> doGetUserAllocatedList(PageRequest request,
                                                       Long roleId, SysUser user, List<Long> deptIds, boolean inRoles) {
-//        Assert.notNull(roleId, "roleId cant null");
-//        List<AggregationOperation> aggregationOperations = new ArrayList<>();
-//
-//        String sys_role_dept_prefix = "sys_dept.";
-//        LookupOperation lookupOperation = LookupOperation.newLookup().
-//                from("sys_dept").
-//                localField("deptId").
-//                foreignField("deptId").
-//                as("sys_dept");
-//        aggregationOperations.add(lookupOperation);
-//
-//
-//        String sys_user_role_prefix = "sys_user_role.";
-//        LookupOperation lookupOperationUserRole = LookupOperation.newLookup().
-//                from("sys_user_role").
-//                localField("userId").
-//                foreignField("userId").
-//                as("sys_user_role");
-//        aggregationOperations.add(lookupOperationUserRole);
-//
-//
-////        String sys_role_prefix = "sys_role.";
-////        LookupOperation lookupOperationSysRole = LookupOperation.newLookup().
-////                from(sys_user_role_prefix + "roleId").
-////                localField(sys_role_menu_prefix + "roleId").
-////                foreignField("roleId").
-////                as("sys_role");
-////        aggregationOperations.add(lookupOperationSysRole);
-//
-//
-//        if (inRoles) {
-//            aggregationOperations.add(match(
-//                    Criteria.where(sys_user_role_prefix + "roleId").is(roleId))
-//            );
-//        } else {
-//            aggregationOperations.add(match(
-//                    Criteria.where(sys_user_role_prefix + "roleId").ne(roleId))
-//            );
-//        }
-//
-//        if (StringUtils.isNotBlank(user.getLoginName())) {
-//            aggregationOperations.add(match(Criteria.where("loginName").regex(user.getLoginName().trim())));
-//        }
-//        if (StringUtils.isNotBlank(user.getPhoneNumber())) {
-//            aggregationOperations.add(match(Criteria.where("phoneNumber").is(user.getPhoneNumber().trim())));
-//        }
-//
-//        //只显示正常
-//        aggregationOperations.add(Aggregation.match(CriteriaUtil.notDeleteCriteria()));
-//
-//        //计算总数量
-//        List<AggregationOperation> aggregationOperationCount = new ArrayList<>();
-//        aggregationOperationCount.addAll(aggregationOperations);
-//        aggregationOperationCount.add(Aggregation.count().as("c"));
-//        Map countMap = mongoTemplate.aggregate(Aggregation.newAggregation(aggregationOperationCount),
-//                SysUser.class, Map.class).getUniqueMappedResult();
-//        long count = countMap == null ? 0 : (Integer) countMap.get("c");
-//
-//        PageRequestUtil.injectAggregationOnlyPage(request,aggregationOperations);
-//
-//        //获取list
-//        List<SysUser> sysUserList = mongoTemplate.aggregate(Aggregation.newAggregation(aggregationOperations),
-//                SysUser.class, SysUser.class).getMappedResults();
-//        return PageModel.instance(count, sysUserList);
 
-//        QSysDept qSysDept = QSysDept.sysDept;
         QSysUser qSysUser = QSysUser.sysUser;
         QSysUserRole qSysUserRole = QSysUserRole.sysUserRole;
         JPAQuery<SysUser> jpa = queryFactory.selectFrom(qSysUser)
@@ -722,7 +574,7 @@ public class SysUserServiceImpl extends CrudBaseServiceImpl<Long,SysUser> implem
      * @param user 用户信息
      */
     @Override
-    public void checkUserAllowed(SysUser user,String actionName) {
+    public void checkAdminModifyAllowed(SysUser user, String actionName) {
         if (!Global.isDebugMode() && user != null && user.getUserId() != null && user.isAdmin()) {
             throw BusinessException.build("不允许" + StringUtils.defaultIfBlank(actionName,"操作") + "超级管理员用户" + StringUtils.defaultIfBlank(user.getLoginName(),""));
         }
